fixed odd physics between rigids and player
fixed issue where an object would be flung if the player is standing on the edge of it. Turns out, this was a Godot Jolt issue. Fixed by setting 'contact speculative distance' and 'contact allowed penetration' to zero.
Also, added a better way for the player to push rigidbodies instead of the hacky 'softpush' way